Share loading screen between editors
