+++ packages/edit-site/src/components/editor/index.js
@@ -9,13 +9,14 @@ import {
 	useCallback,
 } from '@wordpress/element';
 import { useSelect, useDispatch } from '@wordpress/data';
-import { Popover, Button, Notice, Spinner } from '@wordpress/components';
+import { Popover, Button, Notice } from '@wordpress/components';
 import { EntityProvider, store as coreStore } from '@wordpress/core-data';
 import {
 	BlockContextProvider,
 	BlockBreadcrumb,
 	BlockStyles,
 	store as blockEditorStore,
+	LoadingScreen,
 } from '@wordpress/block-editor';
 import {
 	InterfaceSkeleton,
@@ -67,31 +68,6 @@ const interfaceLabels = {
 	/* translators: accessibility text for the navigation sidebar landmark region. */
 	drawer: __( 'Navigation Sidebar' ),
 };
-
-const LoadingScreen = () => (
-	<div
-		style={ {
-			position: 'fixed',
-			top: 0,
-			left: 0,
-			width: '100%',
-			height: '100%',
-			background: '#fff',
-			zIndex: 999999,
-		} }
-	>
-		<div
-			style={ {
-				display: 'flex',
-				alignItems: 'center',
-				justifyContent: 'center',
-				height: '100%',
-			} }
-		>
-			<Spinner style={ { width: 64, height: 64 } } />
-		</div>
-	</div>
-);
 
 function Editor( { onError } ) {
 	const {
